Create image_encryption.py
The encryption tool will swap the red and blue pixel values of each pixel in the image. This is a very basic form of encryption and can be easily broken, but it's a good starting point

How to use

    Save the above code in a file, e.g., image_encryption.py.
    Install the Pillow library if you haven't already: pip install pillow.
    Place the original image in the same directory as the script.
    Run the script: python image_encryption.py.
    The encrypted image will be saved as encrypted_image.png.
    To decrypt the image, run the script again, and the decrypted image will be saved as decrypted_image.png.

Note: This is a very basic form of encryption and should not be used for secure applications. In a real-world scenario, you would want to use a more secure encryption algorithm, such as AES.